delete unnecesary edge functions
